Networking for commands
Now it is ezpz to use a string or similar thing to send a command to a
server and back from a server. This should make things very easy when I
implement the rest of the code. The only significant hiccup I see is
that getting the IP etc. to work between multiple computers may be
difficult at best